backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
langchain-chroma
langgraph
chromadb
openai
pydantic
python-dotenv
lark
IPython